[limits] Compute the budget spend product in 64-bit
hb_budget_spend() evaluated COST * MULT in unsigned int, so the product
wrapped at 2^32 before widening to the int64_t counter.  No current
charge site can reach that (the largest, VARC's axisValues x coords, is
~2.7e8), but this primitive exists to harden against adversarial fonts,
and one cast removes the reasoning burden permanently.

+++ src/hb-limits.hh
@@ -184,7 +184,7 @@
 static HB_ALWAYS_INLINE bool
 hb_budget_spend (int64_t &budget, unsigned int cost, unsigned int mult = 1)
 {
-  budget -= cost * mult;
+  budget -= (int64_t) cost * mult;
   return budget >= 0;
 }
